refactoring: EventCreator is now EventEditor - also the old EventEditor is now deleted
